+++ examples/Eventhdlr/src/event_boundwriting.c
@@ -593,7 +593,7 @@ SCIP_DECL_EVENTEXEC(eventExecBoundwriting)
       /* if no point is found, extend directly */
       if( pch == NULL )
       {
-         (void)SCIPstrncpy(name, eventhdlrdata->filename, (unsigned int)(SCIP_MAXSTRLEN - n));
+         (void)SCIPstrncpy(name, eventhdlrdata->filename, SCIP_MAXSTRLEN - n);
          strncat(name, number, (unsigned int)n);
       }
       else
@@ -605,7 +605,7 @@ SCIP_DECL_EVENTEXEC(eventExecBoundwriting)
          else
             len = (int) (pch-(eventhdlrdata->filename));
 
-         (void)SCIPstrncpy(name, eventhdlrdata->filename, (unsigned int)len);
+         (void)SCIPstrncpy(name, eventhdlrdata->filename, len);
          strncat(name, number, (unsigned int)n);
          assert(len+n < SCIP_MAXSTRLEN);
          name[len+n] = '\0';
